% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{build_keep_args}
\alias{build_keep_args}
\title{Build a list of arguments to keep}
\usage{
build_keep_args(where_to_find_args)
}
\arguments{
\item{where_to_find_args}{}
}
\description{
In the process of building data frames of arguments to \code{FUN},
we need to decide which arguments to keep from each source,
\code{...}, \code{.dat}, and defaults to \code{FUN}.
This function does that work in one place.
}
