% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{build_matsindf_apply_data_frame}
\alias{build_matsindf_apply_data_frame}
\title{Create a data frame consisting of the input data for matsindf_apply()}
\usage{
build_matsindf_apply_data_frame(types, .dat, FUN, ...)
}
\arguments{
\item{types}{The output of \code{matsindf_apply_types()}.}

\item{.dat}{The value of the \code{.dat} argument to \code{matsindf_apply()}, as a list or a data frame.}

\item{dots}{The contents of the \code{...} argument to \code{matsindf_apply()}, as a list or a data frame.}

\item{defaults}{The values of the default arguments for the \code{FUN} argument to \code{matsindf_apply()}.}
}
\value{
A data frame (actually, a \code{tibble})
with columns from \code{dots}, \code{.dat}, and the default values to \code{FUN},
according to precedence rules for \code{matsindf_apply()}..
}
\description{
This is an internal helper function that takes the types list
and creates a data frame from which calculations
can proceed.
}
\details{
This function enforces the precedence rules for \code{matsindf_apply()}, namely that
variables found in \code{...} take priority over
variables found in \code{.dat}, which take priority over
variables found in the default values of \code{FUN}.
}
