% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_to_msg}
\alias{df_to_msg}
\title{Create a message from a data frame}
\usage{
df_to_msg(df)
}
\arguments{
\item{df}{The data frame to be converted to a message}
}
\value{
A string with rows separated by semicolons and entries separated by commas.
}
\description{
This function is especially helpful for cases when a data frame
of missing or unset values is at hand.
Trim unneeded columns, then call this function
to create a string with rows separated by semicolons and entries separated by commas.
}
\examples{
data.frame(a = c(1, 2, 3), b = c("a", "b", "c")) \%>\%
  df_to_msg()
}
