% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{.dat_names_to_keep}
\alias{.dat_names_to_keep}
\title{Tell which \code{.dat} arguments to keep}
\usage{
.dat_names_to_keep(.dat, FUN, ...)
}
\arguments{
\item{.dat}{The \code{.dat} argument to \code{matsindf_apply()}.}

\item{dots}{The \code{...} argument to \code{matsindf_apply()}, as a list.}
}
\value{
A vector of names of \code{.dat} arguments to keep.
}
\description{
\code{.dat} arguments are kept according to this algorithm:
\itemize{
\item If an argument's name in \code{.dat} is also a string value in \code{...}, that argument is kept,
because we will rename and use the argument in \code{.dat}.
\item If an argument's name in \code{.dat} is not in the names of \code{...}, that argument is kept,
because the argument in \code{.dat} will not be replaced by anything in \code{...}.
\item If an argument's name in \code{.dat} is a name in \code{...} but \emph{not} a string value in \code{...},
the argument is deleted from \code{.dat}, because it is not needed
\item If an argument's name in \code{.dat} is a name in \code{...} and also a string value in \code{...}, the argument is kept.
}
}
\details{
For the name of each argument in \code{.dat}, the following table
decides whether that argument should be kept:

\tabular{cccl}{
in \code{names(...)}   \tab  in string values of \code{...}   \tab decision  \tab reason \cr
no                \tab  yes or no                   \tab keep      \tab will not be replaced  \cr
yes or no         \tab  yes                         \tab keep      \tab will be renamed       \cr
yes               \tab  no                          \tab delete    \tab will be replaced      \cr
}
}
\examples{
example_fun <- function(a, b) {c(c = a + b, d = a - b)}
.dat_names_to_keep(.dat = list(a = 2, b = 1, z = 42),
                   FUN = example_fun,
                   a = "z")
}
