% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{group_by_everything_except}
\alias{group_by_everything_except}
\title{Group by all variables except some}
\usage{
group_by_everything_except(.DF, ..., add = FALSE)
}
\arguments{
\item{.DF}{a data frame to be grouped}

\item{...}{a string or list of strings representing column names to be excluded from grouping}

\item{add}{When \code{add = FALSE}, the default, \code{group_by()} will override existing groups.
To add to the existing groups, use \code{add = TRUE}.}
}
\value{
a grouped version of \code{.DF}
}
\description{
This is a convenience function
that allows grouping of a data frame by all variables (columns)
except those variables specified in \code{...}.
}
