% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{handle_empty_data}
\alias{handle_empty_data}
\title{Gracefully handle empty data}
\usage{
handle_empty_data(.dat = NULL, FUN, DF, types)
}
\arguments{
\item{.dat}{The \code{.dat} argument to \code{matsindf_apply()}.}

\item{FUN}{The \code{FUN} argument to \code{matsindf_apply()}.}

\item{DF}{The assembled \code{DF} inside \code{matsindf_apply()}.}

\item{types}{The \code{types} object assembled inside \code{matsindf_apply()}.}
}
\value{
The appropriate return value from \code{matsindf_apply()},
either a zero-length list or a zero-row data frame.
}
\description{
When empty data are provided to \code{matsindf_apply()},
care must be take with the return value.
This function assembles the correct zero-row data frame or
zero-length lists.
}
