% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{handle_null_args}
\alias{handle_null_args}
\title{Gracefully handle \code{NULL} arguments}
\usage{
handle_null_args(.arg)
}
\arguments{
\item{.arg}{One of \code{.dat} or \code{...} (as a list) arguments to \code{matsindf_apply()}.}
}
\value{
A list representation of \code{.arg} with \code{NULL} values handled appropriately.
}
\description{
When \code{NULL} is passed as an element of the \code{.dat} or \code{...} arguments
to \code{matsindf_apply()}, special care must be taken.
This function helps in those situations.
}
