% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_iris}
\alias{long_iris}
\title{Iris data, in a long table}
\format{
A tibble with 600 rows and four variables:
\describe{
\item{flower}{An identifier for the flower being measured}
\item{species}{The species of the flower}
\item{dimension}{The part and dimension of the flower being measured (for example, sepal width or sepal length)}
\item{length}{The measurement, in cm}
}
}
\usage{
long_iris
}
\description{
The Edgar Anderson's iris dataset, which is included in R as "iris", but as a long table.
}
\keyword{datasets}
