\name{matsindf-package}
\alias{matsindf-package}
\alias{matsindf}
\docType{package}
\title{
\packageTitle{matsindf}
}
\description{
\packageDescription{matsindf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{matsindf}
\packageIndices{matsindf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{matsindf}

Maintainer: \packageMaintainer{matsindf}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
}
\seealso{
}
\examples{
}
