% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{matsindf_apply_types}
\alias{matsindf_apply_types}
\title{Determine types of \code{.dat} and \code{...} arguments for matsindf_apply()}
\usage{
matsindf_apply_types(.dat, FUN, ...)
}
\arguments{
\item{.dat}{The \code{.dat} argument to be checked.}

\item{FUN}{The function sent to \code{matsindf_apply()}.}

\item{...}{The list of arguments to be checked.}
}
\value{
A logical list with components named
\code{.dat_null}, \code{.dat_df}, \code{.dat_list}, \code{.dat_names},
\code{FUN_arg_names},
\code{dots_present},
\code{all_dot_num}, \code{all_dots_mats},
\code{all_dots_list}, \code{all_dots_char}, \code{dots_names}, and
\code{arg_source}.
}
\description{
This is a convenience function that returns a list
for the types of \code{.dat} and \code{...} as well as names in \code{.dat} and \code{...},
with components named \code{.dat_null}, \code{.dat_df}, \code{.dat_list}, \code{.dat_names},
\code{FUN_arg_names},
\code{dots_present}, \code{all_dots_num}, \code{all_dots_mats},
\code{all_dots_list}, \code{all_dots_vect}, \code{all_dots_char}, \code{dots_names}, and
\code{arg_source}.
}
\details{
When \code{.dat} is a \code{data.frame}, both \code{.dat_list} and \code{.dat_df} are \code{TRUE}.

When arguments are present in \code{...}, \code{dots_present} is \code{TRUE} but \code{FALSE} otherwise.
When all items in \code{...} are single numbers, \code{all_dots_num} is \code{TRUE} and all other list members are \code{FALSE}.
When all items in \code{...} are matrices, \code{all_dots_mats} is \code{TRUE} and all other list members are \code{FALSE}.
When all items in \code{...} are lists, \code{all_dots_list} is \code{TRUE} and all other list members are \code{FALSE}.
When all items in \code{...} are vectors (including lists), \code{all_dots_vect} is \code{TRUE}.
When all items in \code{...} are character strings, \code{all_dots_char} is \code{TRUE} and all other list members are \code{FALSE}.

\code{arg_source} is a \code{list()} with two named member vectors, \code{.dat} and \code{dots}.
When \code{TRUE}, parameters should be taken from the named location.
Corresponding names will never both be \code{TRUE}.
An error occurs if a parameter needed by \code{FUN} is available from neither \code{.dat} nor \code{...}.
}
\examples{
identity_fun <- function(a, b) {list(a = a, b = b)}
matsindf_apply_types(.dat = NULL, FUN = identity_fun, a = 1, b = 2)
matsindf_apply_types(.dat = data.frame(), FUN = identity_fun,
                     a = matrix(c(1, 2)), b = matrix(c(2, 3)))
matsindf_apply_types(.dat = list(), FUN = identity_fun,
                     a = c(1, 2), b = c(3, 4), c = c(5, 6))
matsindf_apply_types(.dat = NULL, FUN = identity_fun,
                     a = list(1, 2), b = list(3, 4), c = list(5, 6))
matsindf_apply_types(.dat = NULL, FUN = identity_fun,
                     a = "a", b = "b", c = "c")
}
