% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{should_unlist}
\alias{should_unlist}
\title{Tell whether a column can be unlisted}
\usage{
should_unlist(this_col)
}
\arguments{
\item{this_col}{The column to be checked.
Or a \code{data.frame}, in which case every column is checked.}
}
\value{
A boolean. \code{TRUE} if the column can be unlisted, \code{FALSE} otherwise.
When \code{this_col} is a \code{data.frame}, a named boolean vector,
one entry for each column.
}
\description{
When evaluating each row of a data frame in \code{matsindf_apply()},
the result will be a \code{tibble} with list columns.
This function tells whether a column can be unlisted.
}
\examples{
DF <- tibble::tibble(a = list(1, 2, 3), b = c("a", "b", "c"),
                     c = list(matrix(c(42, 43)),
                              matrix(c(44, 45)),
                              matrix(c(46, 47))))
should_unlist(DF$a)
should_unlist(DF$b)
should_unlist(DF$c)
sapply(DF, FUN = function(this_col) {should_unlist(this_col)})
should_unlist(DF)
}
