% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{where_to_get_args}
\alias{where_to_get_args}
\title{Decide where to get each argument to FUN}
\usage{
where_to_get_args(.dat = NULL, FUN, ...)
}
\arguments{
\item{.dat}{The \code{.dat} argument to \code{matsindf_apply()}.}

\item{FUN}{The \code{FUN} argument to \code{matsindf_apply()}.}

\item{...}{The \code{...} argument to \code{matsindf_apply()}.}
}
\value{
A named list wherein the names are the argument names to \code{FUN}.
Values are strings vectors with 2 elements.
The first element is the location at which the named argument should be found,
one of ".dat", "...", or "defaults".
The second element is the variable or argument name that contains the data.
}
\description{
The precedence rules for where to obtain arguments are codified here.
The rules are:
\itemize{
\item Order (highest priority to lowest priority): \code{...}, \code{.dat}, defaults arguments to \code{FUN}.
\item If an element of \code{...} is a single string, the name of that element should be found
in \code{...} (first priority) or defaults to \code{FUN} (second priority)
according to the value.
}
}
\details{
Examples:
\itemize{
\item \code{.dat} has columns \code{a}, \code{b}, and \code{c}.
\code{...} has \code{b = "c"}.
\code{FUN} has arguments \code{a} and \code{b}.
Argument \code{a} of \code{FUN} will be supplied by column \code{a} in \code{.dat}.
Argument \code{b} of \code{FUN} will be supplied by column \code{c} in \code{.dat}.
The return value will be
\code{list(a = c(".dat", "a"), b = c(".dat", "c"))}
\item \code{.dat} has columns \code{b}, and \code{c}.
\code{...} has \code{b = "c"}.
\code{FUN} has arguments \code{a = 1} and \code{b}.
Argument \code{a} of \code{FUN} will be supplied by the default value to \code{FUN} (\code{a = 1}).
Argument \code{b} of \code{FUN} will be supplied by column \code{c} in \code{.dat}.
The return value will be
\code{list(a = c("FUN", "a"), b = c(".dat", "c"))}
}
}
