\name{matsindf-package}
\alias{matsindf-package}
\alias{matsindf}
\docType{package}
\title{
\packageTitle{matsindf}
}
\description{
\packageDescription{matsindf}
}
\details{

# The DESCRIPTION file:
\packageDESCRIPTION{matsindf}
\packageIndices{matsindf}
}
\author{
\packageAuthor{matsindf}
Maintainer: \packageMaintainer{matsindf}
}
\references{
}
\keyword{
}
\seealso{
The
\code{\link[matsindf:matsindf-package]{matsindf}}
package works well with the
\code{\link[byname:byname-package]{byname}}
package.
}
\examples{
}
