% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{should_unlist}
\alias{should_unlist}
\title{Tell whether a column can be unlisted}
\usage{
should_unlist(this_col)
}
\arguments{
\item{this_col}{The column to be checked.
Or a \code{data.frame}, in which case every column is checked.}
}
\value{
A boolean. \code{TRUE} if the column can be unlisted, \code{FALSE} otherwise.
When \code{this_col} is a \code{data.frame}, a named boolean vector,
one entry for each column.
}
\description{
When evaluating each row of a data frame in \code{matsindf_apply()},
the result will be a \code{tibble} with list columns.
This function tells whether a column can be unlisted.
This is internal helper function and should not be called externally.
}
