% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{where_to_get_args}
\alias{where_to_get_args}
\title{Decide where to get each argument to FUN}
\usage{
where_to_get_args(.dat = NULL, FUN, ...)
}
\arguments{
\item{.dat}{The \code{.dat} argument to \code{matsindf_apply()}.}

\item{FUN}{The \code{FUN} argument to \code{matsindf_apply()}.}

\item{...}{The \code{...} argument to \code{matsindf_apply()}.}
}
\value{
A named list wherein the names are the argument names to \code{FUN}.
Values are character vectors with 2 elements.
The first element is named \code{source} and provides
the argument to \code{matsindf_apply()} from which the named argument should be found,
one of ".dat", "FUN", or "...".
The second element is named \code{arg_name} and provides
the variable name or argument name in the source that contains the input data
for the argument to \code{FUN}.
}
\description{
The precedence rules for where to obtain values for the \code{FUN} argument to
\code{matsindf_apply()} are codified here.
The rules are:
\itemize{
\item Precedence order: \code{...}, \code{.dat}, defaults arguments to \code{FUN}
(highest priority to lowest priority).
\item If an element of \code{...} is a character string of length \code{1},
the element of \code{...} provides a mapping between
an item in \code{.dat} (with same name as the value of the character string of length \code{1})
to an argument of \code{FUN} (with the same name as the name of the character string of length \code{1}).
}
}
\details{
Examples:
\itemize{
\item \code{.dat} has columns \code{a}, \code{b}, and \code{c}.
\code{...} has \code{b = "c"}.
\code{FUN} has arguments \code{a} and \code{b}.
Argument \code{a} of \code{FUN} will be supplied by column \code{a} in \code{.dat}.
Argument \code{b} of \code{FUN} will be supplied by column \code{c} in \code{.dat}.
The return value will be
\code{list(a = c(".dat", "a"), b = c(".dat", "c"))}
\item \code{.dat} has columns \code{b}, and \code{c}.
\code{...} has \code{b = "c"}.
\code{FUN} has arguments \code{a = 1} and \code{b}.
Argument \code{a} of \code{FUN} will be supplied by the default value to \code{FUN} (\code{a = 1}).
Argument \code{b} of \code{FUN} will be supplied by column \code{c} in \code{.dat}.
The return value will be
\code{list(a = c("FUN", "a"), b = c(".dat", "c"))}
}
}
